requirements.txt
absl-py==0.10.0
APScheduler==3.6.3
asgiref==3.2.10
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==3.1.1
dj-database-url==0.5.0
Django==3.1.1
django-appconf==1.0.4
django-apscheduler==0.4.2
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.9.2
django-cryptography==1.0
django-tempus-dominus==5.1.2.13
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
httpie==2.2.0
idna==2.10
importlib-metadata==2.0.0
Keras-Preprocessing==1.1.2
Markdown==3.3.2
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.2
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
schedule==0.6.0
six==1.15.0
sqlparse==0.3.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
tzlocal==2.1
urllib3==1.25.10
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.3.1
